Give up on using bytes and just use strings
I was going for conceptual purity here, but in practice I'm always going
to have utf8-encoded branch names, and I want to be able to use regexes.
I _could_ make it work with bytes, but this is project is a lark, and
I'm the only person likely to use it, so let's just do this.