cleaner: save patch ids to file for speed
This is an obvious optimization but is _much_ faster. Instead of
computing patch ids for every commit on main every run, just store them
in a file, and load them from the file so that we only have to run the
diff-tree | patch-id pipeline for the ones we don't already know.

(Also here, some minor Python cleanups I noticed in passing.)