run_git: allow a raw=True kwarg that returns bytes
The code to compute patch ids crashes when faced with a commit that
contains text content that isn't valid utf-8. We can't assume that the
--patch-with-raw output will be valid unicode. (In practice, is almost
always is, especially for me, but Rob has a case where it isn't.)

Fix this by teaching run_git a `raw` keyword argument, and using it when
we run diff-tree. If raw=True, run_git returns a byte string, rather
than an encoded string, and doesn't assume any encoding of its stdin.