cli: add a --no-push argument to avoid pushing
A lot of times I run this I just want to update all my locals, and I
have some junk locally that I don't actually want to push. What I
usually do is re-run with `-i some-branch`, but that's annoying. Fix
this by adding a `--no-push` CLI flag. With it, you'll get a warning
that tells you the branch is newer locally, but we won't actually do
anything.